Correct package name
Because the name of this package is drp_qa, python scripts should
reside not in python/pfs/qa but in python/pfs/drp/qa.

detectorMapQa.py in the old directory is not moved to the new
directory by this commit for fear that the development on-going in
another branch will be troubled.