DM QA auto-scaling
* This add the following config params to the commmand line:

  * `useSigmaRange` default=Truex`
  * `xrange` default 0.1 pix
  * `wrange` default 0.03 nm

If `useSigmaRange` is `True` then the other two params are ignored, otherwise they are set as the upper
and lower limits for the spatial and wavelength residuals respectively.

When run with multiple visits on the command line the sigma range will be shared across the visit, which might
not be desirable as that's comparing across detectors

If the range is larger than 2.5x sigma, the sigma lines will be shown.