Make detectorMapQa workable tickets/PIPE2D-1365 (#3)
This moves all the functionality into the correct location and gets everything working. This includes some basic updates to the plotting. Future 

Note: I'm squashing the commits in this PR because it was not developed according to best-practices and the rebasing has too many conflicts are taking too much time. (Sorry!).